+++ frontend/src/hooks/use-filtered-items.ts
@@ -2,20 +2,16 @@ import Fuse from 'fuse.js';
 import { useEffect, useMemo, useRef } from 'react';
 
 import { FUSE_SEARCH_OPTIONS } from '@/lib/search';
+import {
+  buildTaggedItems,
+  compareItems,
+  type TaggedItem,
+} from '@/lib/tagged-items';
 import { useProfileStore } from '@/stores/profile-store';
 import { type SearchFilterState, useSearchStore } from '@/stores/search-store';
 
 import type { types } from '../../wailsjs/go/models';
-import {
-  type PerPage,
-  type SortDirection,
-  type SortField,
-  type SortState,
-} from '../lib/constants';
-
-export type TaggedItem =
-  | { type: 'mod'; item: types.ModManifest }
-  | { type: 'map'; item: types.MapManifest };
+import { type PerPage, type SortState } from '../lib/constants';
 
 export type { SearchFilterState } from '@/stores/search-store';
 
@@ -98,104 +94,6 @@ export function matchesMapAttributeFilters(
     matchesSingleValueFilter(map.level_of_detail, filters.levelOfDetail) &&
     matchesZeroOrManyValuesFilter(map.special_demand, filters.specialDemand)
   );
-}
-
-export function compareByDirection(
-  a: number,
-  b: number,
-  direction: SortDirection,
-): number {
-  return direction === 'asc' ? a - b : b - a;
-}
-
-export function getTotalDownloads(
-  item: TaggedItem,
-  modDownloadTotals: Record<string, number>,
-  mapDownloadTotals: Record<string, number>,
-): number {
-  return item.type === 'mod'
-    ? (modDownloadTotals[item.item.id] ?? 0)
-    : (mapDownloadTotals[item.item.id] ?? 0);
-}
-
-export function getLastUpdated(item: TaggedItem): number {
-  const timestamp = item.item.last_updated;
-  return typeof timestamp === 'number' && Number.isFinite(timestamp)
-    ? timestamp
-    : 0;
-}
-
-// Helper to determine comparation logic based on sort field and direction
-export function compareItems(
-  a: TaggedItem,
-  b: TaggedItem,
-  sort: SortState,
-  modDownloadTotals: Record<string, number>,
-  mapDownloadTotals: Record<string, number>,
-): number {
-  const compareText = (
-    left: string,
-    right: string,
-    direction: SortDirection,
-  ) =>
-    direction === 'asc' ? left.localeCompare(right) : right.localeCompare(left);
-
-  const compareField = (field: SortField): number => {
-    switch (field) {
-      case 'name':
-        return compareText(
-          a.item.name ?? '',
-          b.item.name ?? '',
-          sort.direction,
-        );
-      case 'country': {
-        const countryA =
-          a.type === 'map' ? ((a.item as types.MapManifest).country ?? '') : '';
-        const countryB =
-          b.type === 'map' ? ((b.item as types.MapManifest).country ?? '') : '';
-        return compareText(countryA, countryB, sort.direction);
-      }
-      case 'author':
-        return compareText(
-          a.item.author ?? '',
-          b.item.author ?? '',
-          sort.direction,
-        );
-      case 'population': {
-        const popA =
-          a.type === 'map'
-            ? ((a.item as types.MapManifest).population ?? 0)
-            : 0;
-        const popB =
-          b.type === 'map'
-            ? ((b.item as types.MapManifest).population ?? 0)
-            : 0;
-        return compareByDirection(popA, popB, sort.direction);
-      }
-      case 'downloads': {
-        const downloadsA = getTotalDownloads(
-          a,
-          modDownloadTotals,
-          mapDownloadTotals,
-        );
-        const downloadsB = getTotalDownloads(
-          b,
-          modDownloadTotals,
-          mapDownloadTotals,
-        );
-        return compareByDirection(downloadsA, downloadsB, sort.direction);
-      }
-      case 'last_updated': {
-        const updatedA = getLastUpdated(a);
-        const updatedB = getLastUpdated(b);
-        return compareByDirection(updatedA, updatedB, sort.direction);
-      }
-      default:
-        return 0;
-    }
-  };
-
-  return compareField(sort.field);
 }
 
 // Seeded hash function to provide consistent "random" sorting. Stable across renders, but different across sessions
@@ -302,17 +200,7 @@ export function useFilteredItems({
     setPage(1);
   }, [filters, setPage]);
 
-  const allItems = useMemo<TaggedItem[]>(() => {
-    const modItems: TaggedItem[] = (mods || []).map((m) => ({
-      type: 'mod' as const,
-      item: m,
-    }));
-    const mapItems: TaggedItem[] = (maps || []).map((m) => ({
-      type: 'map' as const,
-      item: m,
-    }));
-    return [...modItems, ...mapItems];
-  }, [mods, maps]);
+  const allItems = useMemo<TaggedItem[]>(() => buildTaggedItems(mods, maps), [mods, maps]);
 
   const filtered = useMemo(() => {
     return filterAndSortTaggedItems(
